Improves Scanning and Implements parsing and codegen of multiple functions
The scanner can now correctly scan most of the keywords and operators
that i anticipate i'll need for river.
There are still a couple edge cases and special scanning features
remaining, such as differenciating between floating point and double
values, but those can come later.
The scanner still needs to be rewritten to be a little less messy, but
that is for another day.

The parser and codegen modules can now scan through and assemble a file
with multiple function definitions, which is pretty coool.

Error handling, Error reporting, and memory management all leave a ton
to be desired though, as always... those will come soon.